Print a message to the user when doing an up-thread leaves an empty thread.
When the user does an up thread and the resulting merge leaves a tree
with no conflicts and a tree that does not differ from the one that
it just came from it means that the thread it has just moved to
is "empty". This can indicate that the change that was being developed in
the thread has been fully merged upstream, and so the user may wish to
combine the thread. We just print a message at this stage so that the
user is aware of this. We may want to add an option to automatically
combine the thread later.